// src/modules/event/common/event.enums.ts
export enum StreakSize {
  HUNDRED = 100,
  NINETY = 90,
  SEVENTY = 70,
  SIXTY = 60,
  FORTY = 40,
  THIRTY = 30,
  TEN = 10
}

export enum AttendanceStreakReward {
  HUNDRED = 50,
  THIRTY = 10,
  TEN = 3,
}

export enum EventType {
  ATTENDANCE = 'attendance',
  PARTICIPATION = 'participation',
}
export enum ParticipationStreakReward {
  HUNDRED = 100,
  THIRTY = 50,
  TEN = 5,
}